Переименовал файл скрипта
